+++ lib/airbrake/rails/action_controller_performance_breakdown_subscriber.rb
@@ -10,17 +10,28 @@ module Airbrake
         payload = event.payload
 
         routes.each do |route, method|
+          next if (groups = build_groups(payload)).none?
+
           Airbrake.notify_performance_breakdown(
             method: method,
             route: route,
             response_type: payload[:format],
-            groups: {
-              db: payload[:db_runtime] || 0,
-              view: payload[:view_runtime] || 0
-            },
+            groups: groups,
             start_time: event.time
           )
         end
+      end
+
+      def build_groups(payload)
+        groups = {}
+
+        db_runtime = payload[:db_runtime] || 0
+        groups[:db] = db_runtime if db_runtime > 0
+
+        view_runtime = payload[:view_runtime] || 0
+        groups[:view] = view_runtime if view_runtime > 0
+
+        groups
       end
     end
   end