+++ spec/apps/rails/dummy_app.rb
@@ -30,6 +30,7 @@ class DummyApp < Rails::Application
   routes.append do
     get '/' => 'dummy#index'
     get '/crash' => 'dummy#crash'
+    get '/breakdown' => 'dummy#breakdown'
     get '/notify_airbrake_helper' => 'dummy#notify_airbrake_helper'
     get '/notify_airbrake_sync_helper' => 'dummy#notify_airbrake_sync_helper'
     get '/active_record_after_commit' => 'dummy#active_record_after_commit'
@@ -100,7 +101,8 @@ class DummyController < ActionController::Base
       'dummy/active_record_after_rollback.html.erb' => 'active_record_after_rollback',
       'dummy/active_job.html.erb' => 'active_job',
       'dummy/resque.html.erb' => 'resque',
-      'dummy/delayed_job.html.erb' => 'delayed_job'
+      'dummy/delayed_job.html.erb' => 'delayed_job',
+      'dummy/breakdown.html.erb' => 'breakdown'
     )
   ]
 
@@ -109,6 +111,11 @@ class DummyController < ActionController::Base
   def crash
     Book.create(title: 'book')
     raise AirbrakeTestError
+  end
+
+  def breakdown
+    Book.create(title: 'breakdown')
+    Book.all
   end
 
   def notify_airbrake_helper